add force close option to close_channel RPC + init ln dir
